requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
Deprecated==1.2.13
Flask==2.2.2
Flask-Limiter==3.1.0
idna==3.4
instagrapi==1.16.30
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.1.6
MarkupSafe==2.1.1
ordered-set==4.1.0
packaging==23.0
Pillow==9.4.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.5
pycryptodomex==3.15.0
pydantic==1.10.2
Pygments==2.14.0
PySocks==1.7.1
python-decouple==3.7
requests==2.28.2
rich==12.6.0
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.14.1
